feat: connecting to the Auction
This shows that we can establish a connection from the Sniper to the
auction, which means that we had to sort out details such as
interpreting the item and user credentials from the command-line
arguments and using the Smack library.

We're leaving the message contents until later because we only have one
message type, so sending an empty value is enough to prove the
connection.