feat: unpacking a price message
We write a rough implementation to prove that we know how to make the
code do what's required and then we REFACTOR.

That "first-cut" code is NOT finished. It's good enough to sort out our
ideas and make sure we have everything in place, but it's unlikely to
express its intensions cleanly.